Win32 build

+++ examples/fft_benchmark.cpp
@@ -186,7 +186,7 @@ int main(int argc, char **argv) {
 
     if(composite != 0) {
         for(auto n = twos.begin() ; n != twos.end() ; n++) {
-            const size_t m = *n / composite + 0.5;
+            const size_t m = static_cast<size_t>(*n / composite + 0.5);
             if(m > 1) ns.push_back(composite * m);
         }
     }
